Feature (entrypoint): Add azure-pipelines `build.yml` and `test.yml` entrypoint templates for improved step control
`build` and `test` steps used to be coupled with `continuous-build.yml` template. Adding `build.yml` and `test.yml` allows execution of only steps applicable to them, allowing for more efficient use of CI resources. For example, the `continuous-build.yml` entrypoint has all along incorporated `build` and `test` steps. The latter is unneeded for publishing of modules but had to be run due to a lack of a `build` entrypoint other than `continuous-build.yml`. With the addition `build.yml` entrypoints, `test` steps can now be skipped in publishing processes.